+++ tests/generic/baker_recipes.py
@@ -1,17 +1,16 @@
 # ATTENTION: Recipes defined for testing purposes only
+from datetime import timedelta
 from decimal import Decimal
-from model_bakery.recipe import Recipe, foreign_key, seq
-from model_bakery.recipe import related
+
+from model_bakery.recipe import Recipe, foreign_key, related, seq
 from model_bakery.timezone import now
 from tests.generic.models import (
     TEST_TIME,
-    Person,
     Dog,
     DummyDefaultFieldsModel,
     DummyUniqueIntegerFieldModel,
+    Person,
 )
-
-from datetime import timedelta
 
 person = Recipe(
     Person,