+++ tests/test_extending_bakery.py
@@ -1,8 +1,7 @@
 import pytest
-
 from model_bakery import baker
-from model_bakery.random_gen import gen_from_list
 from model_bakery.exceptions import CustomBakerNotFound, InvalidCustomBaker
+from model_bakery.random_gen import gen_from_list
 from tests.generic.models import Person